requirements.txt
pytest==9.0.0
selene == 2.0.0rc9
allure-pytest==2.15.0
allure-python-commons
selenium==4.38.0
python-dotenv==1.2.01